+++ src/main/java/com/alldeals/domain/BlogCategory.java
@@ -16,14 +16,14 @@ public class BlogCategory implements Serializable{
 	
 
 	@Id
-	@GeneratedValue(strategy=GenerationType.AUTO)
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
     private int blogCategoryId;
 
 	@NotBlank @Size(min = 3, max=12,message="{Size.blogcategory}")
     @Column(name = "blog_cat_name")
     private String blogCategoryName;
 
-    private String description;
+
     @OneToMany(mappedBy = "blogCategory", cascade = CascadeType.ALL,
             fetch = FetchType.EAGER, orphanRemoval = true)
     private List<BlogArticle> articles = new ArrayList<>();
@@ -51,13 +51,7 @@ public class BlogCategory implements Serializable{
         this.blogCategoryName = blogCategoryName;
     }
 
-    public String getDescription() {
-        return description;
-    }
 
-    public void setDescription(String description) {
-        this.description = description;
-    }
 
     public List<BlogArticle> getArticles() {
         return articles;
