+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 jdbc.driverClassName = com.mysql.jdbc.Driver
-jdbc.url = jdbc:mysql://localhost:3308/alldeals
+jdbc.url = jdbc:mysql://localhost:3306/alldeals
 jdbc.username = root
 jdbc.password =
 hibernate.hbm2ddl.auto=create
