+++ src/main/resources/messages.properties
@@ -59,13 +59,13 @@ description = Description
 #BLOG&CATEGOR
 blog.title = New Blog
 label.title = Blog title
+label.blogDescription = Description
 label.content = Blog Content
 label.blogCategory = Blog Category
 label.publishDate = Publishing Date
 label.blogPhoto = Blog picture
 category.title = New Blog Category
-label.blogCategoryName = Category name
-Save =Save
+Save = Save
 
 
 #BLOG&CATEGORY ERRORS
