+++ src/main/webapp/WEB-INF/layouts/template/navigation.jsp
@@ -23,7 +23,7 @@
                     <a href="<spring:url value="/market/products"/>" class="nav-link">Deal Alert</a>
                 </li>
                 <li class="nav-item">
-                    <a href="<spring:url value="/createBlog"/>" class="nav-link">Blog</a>
+                    <a href="<spring:url value="/blogList"/>" class="nav-link">Blog</a>
                 </li>
             </ul>
             <form class="form-inline mt-2 mt-md-0 mr-2">
